+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+  "cSpell.language": "en,en-GB",
   "cSpell.words": [
     "Dalek",
     "Faceit",
